+++ tests/index.js
@@ -10,6 +10,7 @@ import * as customCompare from './custom-compare.test.js'
 import * as inlineNodes from './inline-nodes.test.js'
 import * as attrAttribution from './attr-attribution.test.js'
 import * as brokenSchema from './broken-schema.test.js'
+import * as ySyncRdt from './y-sync-rdt.test.js'
 import * as prosemirrorRdt from './prosemirror-rdt.test.js'
 import * as swallowFormats from './swallow-formats.test.js'
 // import * as tr from './tr.test.js'
@@ -33,6 +34,7 @@ runTests({
   inlineNodes,
   attrAttribution,
   brokenSchema,
+  ySyncRdt,
   prosemirrorRdt,
   swallowFormats
   // prosemirror,